+++ src/main/scala/lasp/hapi/service/CapabilitiesService.scala
@@ -6,17 +6,15 @@ import org.http4s.HttpService
 import org.http4s.circe._
 import org.http4s.dsl.Http4sDsl
 
-import lasp.hapi.server.HapiServer
-
 /** Implements the `/capabilities` endpoint. */
 class CapabilitiesService[F[_]: Effect] extends Http4sDsl[F] {
 
   val service: HttpService[F] =
     HttpService[F] {
-      case GET -> Root / "capabilities" =>
+      case GET -> Root / "hapi" / "capabilities" =>
         Ok(
           Capabilities(
-            HapiServer.version,
+            HapiService.version,
             Status.`1200`,
             List("binary", "csv", "json")
           ).asJson
