+++ src/main/scala/lasp/hapi/service/CatalogService.scala
@@ -9,7 +9,7 @@ class CatalogService[F[_]: Effect] extends Http4sDsl[F] {
 
   val service: HttpService[F] =
     HttpService[F] {
-      case GET -> Root / "catalog" =>
+      case GET -> Root / "hapi" / "catalog" =>
         Ok("Hello from HAPI!")
     }
 }
