+++ src/main/scala/lasp/hapi/service/HapiService.scala
@@ -1,50 +1,53 @@
 package lasp.hapi.service
 
+import scala.concurrent.duration._
+
 import cats.effect.Effect
 import cats.implicits._
 import org.http4s.HttpService
-import org.http4s.MediaType.`text/html`
-import org.http4s.dsl.Http4sDsl
-import org.http4s.headers.`Content-Type`
-import scalatags.Text.all._
+import org.http4s.server.middleware.CORS
+import org.http4s.server.middleware.CORSConfig
 
-/** Implements the `/hapi` endpoint. */
-class HapiService[F[_]: Effect] extends Http4sDsl[F] {
+/**
+ * A grouping of all four required HAPI endpoints.
+ *
+ * The HAPI spec defines four required endpoints. This class groups
+ * those four endpoints and exposes them as a single service that
+ * implements the HAPI spec.
+ *
+ * These are the routes for each service:
+ *
+ *  - `/hapi/capabilities`
+ *  - `/hapi/catalog`
+ *  - `/hapi/data`
+ *  - `/hapi/info`
+ */
+class HapiService[F[_]: Effect] {
 
-  val landingPage: Frag =
-    html(
-      body(
-        h1("LASP HAPI Server"),
-        p("""This server supports the HAPI 2.0 specification for
-        |delivery of time series data. The server consists of the
-        |following 4 REST-like endpoints that will respond to HTTP GET
-        |requests:""".stripMargin.replaceAll("\n", " ")),
-        ul(
-          li(
-            a(href := "hapi/capabilities")("capabilities"),
-            " - list the output formats the server can emit"
-          ),
-          li(
-            a(href := "hapi/catalog")("catalog"),
-            " - list the datasets that are available"
-          ),
-          li(
-            a(href := "hapi/info")("info"),
-            " - describe a dataset"
-          ),
-          li(
-            a(href := "hapi/data")("data"),
-            " - stream content of a dataset"
-          )
-        )
-      )
-    )
+  private val corsConfig: CORSConfig = CORSConfig(
+    anyOrigin        = true,
+    anyMethod        = false,
+    allowedMethods   = Set("GET").some,
+    allowedHeaders   = Set("Content-Type").some,
+    allowCredentials = false,
+    maxAge           = 1.day.toSeconds
+  )
 
-  val service: HttpService[F] =
-    HttpService[F] {
-      case GET -> Root / "hapi" =>
-        Ok(landingPage.render).map {
-          _.withContentType(`Content-Type`(`text/html`))
-        }
+  /** A service composed of all four required HAPI endpoints. */
+  val service: HttpService[F] = {
+    // If you see a red squiggly here it's probably a lie.
+    val service = {
+      new CapabilitiesService[F].service <+>
+      new InfoService[F].service         <+>
+      new CatalogService[F].service      <+>
+      new DataService[F].service
     }
+    CORS(service, corsConfig)
+  }
+}
+
+object HapiService {
+
+  /** Version of HAPI spec. */
+  val version: String = "2.0"
 }
